+++ resource/src/set.rs
@@ -1,5 +1,9 @@
 use {
-    crate::{descriptor, escape::Handle},
+    crate::{
+        descriptor,
+        escape::Handle,
+        util::{device_owned, Device, DeviceId},
+    },
     gfx_hal::{pso::DescriptorSetLayoutBinding, Backend, Device as _},
     relevant::Relevant,
     smallvec::SmallVec,
@@ -7,23 +11,28 @@ use {
 
 /// Descriptor set layout info.
 #[derive(Clone, Debug)]
-pub struct Info {
-    /// Bindings
+pub struct DescriptroSetInfo {
+    /// Bindings.
     pub bindings: Vec<DescriptorSetLayoutBinding>,
 }
 
-impl Info {
+impl DescriptroSetInfo {
+    /// Get descriptor ranges of the layout.
     pub fn ranges(&self) -> descriptor::DescriptorRanges {
         descriptor::DescriptorRanges::from_bindings(&self.bindings)
     }
 }
 
+/// Generic descriptor set layout resource wrapper.
 #[derive(Debug)]
 pub struct DescriptorSetLayout<B: Backend> {
+    device: DeviceId,
     raw: B::DescriptorSetLayout,
-    info: Info,
+    info: DescriptroSetInfo,
     relevant: Relevant,
 }
+
+device_owned!(DescriptorSetLayout<B>);
 
 impl<B> DescriptorSetLayout<B>
 where
@@ -31,44 +40,53 @@ where
 {
     /// Create new descriptor set layout
     pub unsafe fn create(
-        device: &B::Device,
-        info: Info,
+        device: &Device<B>,
+        info: DescriptroSetInfo,
     ) -> Result<Self, gfx_hal::device::OutOfMemory> {
         let raw = device
             .create_descriptor_set_layout(&info.bindings, std::iter::empty::<B::Sampler>())?;
 
         Ok(DescriptorSetLayout {
+            device: device.id(),
             raw,
             info,
             relevant: Relevant,
         })
     }
 
-    pub unsafe fn dispose(self, device: &B::Device) {
+    /// Destroy descriptor set layout resource.
+    pub unsafe fn dispose(self, device: &Device<B>) {
+        self.assert_device_owner(device);
         device.destroy_descriptor_set_layout(self.raw);
         self.relevant.dispose();
     }
 
+    /// Get reference to raw descriptor set layout resource.
     pub fn raw(&self) -> &B::DescriptorSetLayout {
         &self.raw
     }
 
+    /// Get mutable reference to raw descriptor set layout resource.
     pub unsafe fn raw_mut(&mut self) -> &mut B::DescriptorSetLayout {
         &mut self.raw
     }
 
-    pub fn info(&self) -> &Info {
+    /// Get descriptor set layout info.
+    pub fn info(&self) -> &DescriptroSetInfo {
         &self.info
     }
 }
 
-/// Descriptor set object wrapper.
+/// Generic descriptor set resource wrapper.
 #[derive(Debug)]
 pub struct DescriptorSet<B: Backend> {
+    device: DeviceId,
     set: descriptor::DescriptorSet<B>,
     layout: Handle<DescriptorSetLayout<B>>,
     relevant: Relevant,
 }
+
+device_owned!(DescriptorSet<B>);
 
 impl<B> DescriptorSet<B>
 where
@@ -76,7 +94,7 @@ where
 {
     /// Create new descriptor set.
     pub unsafe fn create(
-        device: &B::Device,
+        device: &Device<B>,
         allocator: &mut descriptor::DescriptorAllocator<B>,
         layout: Handle<DescriptorSetLayout<B>>,
     ) -> Result<Self, gfx_hal::device::OutOfMemory> {
@@ -86,6 +104,7 @@ where
 
         assert_eq!(sets.len() as u32, 1);
         Ok(DescriptorSet {
+            device: device.id(),
             set: sets.swap_remove(0),
             layout: layout.clone(),
             relevant: Relevant,
@@ -94,7 +113,7 @@ where
 
     /// Create new descriptor sets.
     pub unsafe fn create_many(
-        device: &B::Device,
+        device: &Device<B>,
         allocator: &mut descriptor::DescriptorAllocator<B>,
         layout: Handle<DescriptorSetLayout<B>>,
         count: u32,
@@ -113,6 +132,7 @@ where
         assert_eq!(sets.len() as u32, count);
 
         extend.extend(sets.into_iter().map(|set| DescriptorSet {
+            device: device.id(),
             set,
             layout: layout.clone(),
             relevant: Relevant,
@@ -121,20 +141,24 @@ where
         Ok(())
     }
 
+    /// Destroy descriptor set resource.
     pub unsafe fn dispose(self, allocator: &mut descriptor::DescriptorAllocator<B>) {
         allocator.free(Some(self.set));
         self.relevant.dispose();
     }
 
+    /// Get reference to raw descriptor set resource.
     pub fn raw(&self) -> &B::DescriptorSet {
         self.set.raw()
     }
 
+    /// Get mutable reference to raw descriptor set resource.
     pub unsafe fn raw_mut(&mut self) -> &mut B::DescriptorSet {
         self.set.raw_mut()
     }
 
-    pub fn layout(&mut self) -> &DescriptorSetLayout<B> {
+    /// Get layout of descriptor set.
+    pub fn layout(&mut self) -> &Handle<DescriptorSetLayout<B>> {
         &self.layout
     }
 }
